content-security-policy/base-uri/base-uri_iframe_sandbox.sub.html

<!DOCTYPE HTML>
<html>

<head>
    <title>base-uri works correctly inside a sandboxed iframe.</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>

    <!-- CSP served: base-uri 'self' -->
</head>

<body>
    <h1>base-uri works correctly inside a sandboxed iframe.</h1>
    <div id='log'></div>

    <script>
        window.addEventListener('securitypolicyviolation', function(e) {
            assert_unreached('No CSP violation report has fired.');
        });

        async_test(function(t) {
            window.addEventListener('message', t.step_func(function(e) {
                if (e.data.startsWith('SAME-ORIGIN,')) {
                    assert_equals(e.data, 'SAME-ORIGIN,' + location.origin + '/');
                    t.done();
                }
            }));
            var e = document.createElement('iframe');
            e.sandbox = 'allow-scripts';
            e.style.display = 'none';
            e.srcdoc = `
            <script>
              window.addEventListener('securitypolicyviolation', function() {
                top.postMessage('SAME-ORIGIN,FAIL', '*');
              });
            </sc` + `ript>
            <base href="{{location[scheme]}}://{{domains[]}}:{{ports[http][0]}}/">
            <script>
              top.postMessage('SAME-ORIGIN,' + document.baseURI, '*');
            </sc` + `ript>`;
            document.body.appendChild(e);
        }, 'base-uri \'self\' works with same-origin sandboxed iframes.');

        async_test(function(t) {
            window.addEventListener('message', t.step_func(function(e) {
                console.log(e);
                if (e.data.startsWith('FOREIGN-ORIGIN')) {
                    assert_equals(e.data, 'FOREIGN-ORIGIN,' +
                        location.origin + '/');
                    t.done();
                }
            }));
            var e = document.createElement('iframe');
            e.sandbox = 'allow-scripts';
            e.style.display = 'none';
            e.srcdoc = `
            <script>
              window.addEventListener('securitypolicyviolation',
                function(violation) {
                  console.log(violation);
                  if (violation.blockedURI !== '{{location[scheme]}}://{{domains[www2]}}:{{ports[http][0]}}/') {
                      return;
                  }
                  assert_equals(violation.effectiveDirective, 'base-uri');
                  top.postMessage('FOREIGN-ORIGIN,' + document.baseURI, '*');
              });
            </sc` + `ript>
            <base href="{{location[scheme]}}://{{domains[www2]}}:{{ports[http][0]}}/">
            <script>
              top.postMessage('FOREIGN-ORIGIN,' + document.baseURI, '*');
            </sc` + `ript>`;
            document.body.appendChild(e);
        }, 'base-uri \'self\' blocks foreign-origin sandboxed iframes.');
    </script>

</body>

</html>
